requirements.txt
anthropic
pytest
requests
pyyaml
robotframework
pytest-json-report
